+++ noodles-bam/src/lazy/record.rs
@@ -424,7 +424,7 @@ impl TryFrom<Record> for sam::alignment::RecordBuf {
         builder = builder.set_template_length(template_length);
 
         builder = builder
-            .set_sequence(lazy_record.sequence().try_into()?)
+            .set_sequence(lazy_record.sequence().into())
             .set_quality_scores(lazy_record.quality_scores().into())
             .set_data(lazy_record.data().try_into()?);