+++ noodles-sam/src/reader/record/sequence.rs
@@ -1,50 +1,31 @@
-use std::{error, fmt, mem};
+use std::{error, fmt};
 
-use crate::record::{
-    sequence::{base, Base},
-    Sequence,
-};
+use crate::alignment::record_buf::Sequence;
 
 /// An error when a raw SAM record sequence fails to parse.
 #[derive(Clone, Debug, Eq, PartialEq)]
 pub enum ParseError {
     /// The input is empty.
     Empty,
-    /// A base is invalid.
-    InvalidBase(base::TryFromCharError),
 }
 
-impl error::Error for ParseError {
-    fn source(&self) -> Option<&(dyn error::Error + 'static)> {
-        match self {
-            Self::Empty => None,
-            Self::InvalidBase(e) => Some(e),
-        }
-    }
-}
+impl error::Error for ParseError {}
 
 impl fmt::Display for ParseError {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         match self {
             Self::Empty => write!(f, "empty input"),
-            Self::InvalidBase(_) => write!(f, "invalid base"),
         }
     }
 }
 
-pub(crate) fn parse_sequence(src: &[u8], sequence: &mut Sequence) -> Result<(), ParseError> {
+pub(super) fn parse_sequence(src: &[u8], sequence: &mut Sequence) -> Result<(), ParseError> {
     if src.is_empty() {
         return Err(ParseError::Empty);
     }
 
-    let mut bases = Vec::from(mem::take(sequence));
-
-    for &n in src {
-        let base = Base::try_from(n).map_err(ParseError::InvalidBase)?;
-        bases.push(base);
-    }
-
-    *sequence = Sequence::from(bases);
+    let sequence = sequence.as_mut();
+    sequence.extend(src);
 
     Ok(())
 }
@@ -57,19 +38,18 @@ mod tests {
     fn test_parse_sequence() -> Result<(), ParseError> {
         let mut sequence = Sequence::default();
 
-        sequence.clear();
+        sequence.as_mut().clear();
         parse_sequence(b"ACGT", &mut sequence)?;
-        let expected = Sequence::from(vec![Base::A, Base::C, Base::G, Base::T]);
+        let expected = Sequence::from(b"ACGT".to_vec());
         assert_eq!(sequence, expected);
 
-        sequence.clear();
-        assert_eq!(parse_sequence(b"", &mut sequence), Err(ParseError::Empty));
+        sequence.as_mut().clear();
+        parse_sequence(&[0x07], &mut sequence)?;
+        let expected = Sequence::from(vec![0x07]);
+        assert_eq!(sequence, expected);
 
-        sequence.clear();
-        assert!(matches!(
-            parse_sequence(&[0x07], &mut sequence),
-            Err(ParseError::InvalidBase(_))
-        ));
+        sequence.as_mut().clear();
+        assert_eq!(parse_sequence(b"", &mut sequence), Err(ParseError::Empty));
 
         Ok(())
     }